test(cli): split transpile tests
